node: add overlayscrollbars
Mainly doing this because Mozilla, of course, has fallen behind on implementing things that every other browser (including Safari!) already has. There is not enough skinning support for the scrollbar at the time of writing this, so I have to rely on a third-party package to do the work Mozilla refuses to do, sadly.