pages/game: empty this page
This page was originally for testing out layout and components related to game specific pages, but since they've been hooked up to a data source, this is no longer needed. However, it will eventually be repurpoused to show the games list, and search results, but just not at the moment. Why am I doing this? Webpack can't build this page, likely because the components are edited to take props, and those props are undefined at build time when Next.js is trying to make the page static. This is completely valid behavior, but I don't want to delete the page since it'll be used later, so for now, it will be empty.