components/game/gameReviews: delete load barring mechanism
The only purpose it seem to be serving is complicating control flow. There is no way to really tell if it was working, but it didn't seem helpful enough to keep around. And, when removing it, responsiveness and behavior improved. So, I just went ahead and removed it, since it only seemed to be painful keeping it around. I suppose I could come back and regret this choice, but somehow, I doubt that.