no short summary, read long message/description
Git is one of those retarded overcomplicated tools that will sometimes get rid of work you've done, simply because you wanted to rewrite the history to fix one of Git's design mistakes. One of the dumbest decisions by the Git team however, has to be discarding commits that are STASHED mind you, and then NOT restoring them when I pop the stash. Seriously, like, what the fuck? Why would it ever do a destructive action when trying to backup work? I can't see why it works the way it does, but it's one of those things that again, makes me wonder where we went wrong when we decided to engineer software. The flip side is that while the stash might have permanently destroyed pertinent and potentially useful commit information, I still kept the most useful bits, which is the diffs. So at least the engineers aren't stupid enough to erase these.

Anyways, I'm not going to bother try to reassemble the context and summaries when I wrote those commits. Git has so many places where it's poorly designed, so I don't feel the need to redo those when it should have kept it in the first place. The work I did simply boils down to this: make a functional screenshot page, and have the carousel link to those pages. In the middle I might have extracted out a helper utility function into it's own file, but that's a much smaller part of the changes I made. Prayerfully, Git doesn't make the same mistake again, and this idotic design defects get patches out of crucial source control infrastructure software.